end first step of refactoring
